refactor: 상수 클래스 생성 및 상수 이동
상수 클래스 생성하여 상수 이동, Lotto Class 이름 변경 -> LottoMachine
