+++ src/main/java/lotto/common/LottoConstants.java
@@ -1,0 +1,14 @@
+package lotto.common;
+
+public class LottoConstants {
+    public static final int ZERO = 0;
+    public static final int LOTTO_PRICE = 1000;
+    public static final int MIN_LOTTO_NUMBER = 1;
+    public static final int MAX_LOTTO_NUMBER = 45;
+    public static final int LOTTO_NUMBER_COUNT = 6;
+    public static final int PRINCIPAL_RATE = 1;
+    public static final String SPLIT_SYMBOL = ",";
+    public static final String WINNING_NUMBER_CREATE_FAIL_MESSAGE = "당첨 번호 생성에 실패 했습니다.";
+    public static final String LOTTO_NUMBER_OUT_OF_BOUND_MESSAGE = "로또 번호는 1 ~ 45 까지 입니다.";
+    public static final String WRONG_INPUT_NUMBER = "번호를 잘 못 입력하셨습니다.";
+}
