+++ src/main/java/lotto/model/LottoResult.java
@@ -1,5 +1,7 @@
 package lotto.model;
 
+import lotto.view.LottoResultMessage;
+
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
