+++ src/test/java/lotto/LottoMachineNumberTest.java
@@ -4,11 +4,11 @@ import lotto.model.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static lotto.model.LottoNumber.LOTTO_NUMBER_OUT_OF_BOUND_MESSAGE;
+import static lotto.common.LottoConstants.LOTTO_NUMBER_OUT_OF_BOUND_MESSAGE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-public class LottoNumberTest {
+public class LottoMachineNumberTest {
 
     @Test
     @DisplayName("로또 번호 객체 생성")
