+++ src/test/java/lotto/LottoMachineTest.java
@@ -1,24 +1,24 @@
 package lotto;
 
-import lotto.model.Lotto;
+import lotto.model.LottoMachine;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class LottoTest {
+public class LottoMachineTest {
 
-    private Lotto lotto;
+    private LottoMachine lottoMachine;
 
     @BeforeEach
     public void setup() {
-        lotto = new Lotto();
+        lottoMachine = new LottoMachine();
     }
 
     @Test
     @DisplayName("입력 받은 금액 / 1000 만큼 구매")
     public void buyLottoCount(){
-        assertThat(lotto.buyCount(15000)).isEqualTo(15);
+        assertThat(lottoMachine.buyCount(15000)).isEqualTo(15);
     }
 }
