+++ src/test/java/lotto/WinningNumbersTest.java
@@ -4,8 +4,8 @@ import lotto.model.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static lotto.model.LottoNumber.LOTTO_NUMBER_OUT_OF_BOUND_MESSAGE;
-import static lotto.model.WinningNumbers.WINNING_NUMBER_CREATE_FAIL_MESSAGE;
+import static lotto.common.LottoConstants.LOTTO_NUMBER_OUT_OF_BOUND_MESSAGE;
+import static lotto.common.LottoConstants.WINNING_NUMBER_CREATE_FAIL_MESSAGE;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 public class WinningNumbersTest {
